+++ src/anemoi/inference/plugin.py
@@ -94,11 +94,15 @@ class AIModelPlugin(Model):
         return PluginRunner(self._checkpoint, device=self.device)
 
     def run(self):
+
         if self.deterministic:
             self.torch_deterministic_mode()
 
-        input = FieldListInput(self.runner, input_fields=self.all_fields)
-        output = CallbackOutput(self.runner, write=self.write)
+        input_kwargs = self.input.anemoi_plugin_input_kwargs()
+        output_kwargs = self.input.anemoi_plugin_input_kwargs()
+
+        input = FieldListInput(self.runner, input_fields=self.all_fields, **input_kwargs)
+        output = CallbackOutput(self.runner, write=self.write, **output_kwargs)
 
         input_state = input.create_input_state(date=self.start_datetime)