+++ CHANGELOG.md
@@ -89,6 +89,8 @@ Keep it human-readable, your future self will thank you!
 - Add support for huggingface
 - Add support for `output_frequency` to write less output
 - Added ability to run inference over multiple GPUs [#55](https://github.com/ecmwf/anemoi-inference/pull/55)
+- Add `processor` class
+- Revisit GRIB encoding
 
 ### Changed
 - Change `write_initial_state` default value to `true`
@@ -104,6 +106,7 @@ Keep it human-readable, your future self will thank you!
 - Fix `__version__` import in init
 - use earthkit-data 0.11.2
 - Fix SimpleRunner
+- Better plots when fields have missing values
 
 ### Removed
 - ci: turn off hpc workflow